refactor(project): Initial build system
Organized project structure

- Source files and Header files are now separated (src and include
  folders, respectively)

- Tests are now separated from data structure implementation.

Starts to implement cmake build system
