+++ include/GPT-originals/uarray.h
