+++ rank-test.cpp
@@ -1,1 +1,1 @@
-#include "jacobsonrank.h"
+#include "include/jacobsonrank.h"
