--- src/FUN/general_tree_2n.c
@@ -1,7 +1,0 @@
-// Create a program that writes (and prints using the
-// Graphviz library) general trees as a 2n bit array
-// and  builds up a general tree from that array
-
-// Utils: 
-//     "https://graphviz.org/documentation/",
-//     "General-Tree-Encode-2n-bits.jpg",
