+++ src/GPT-originals/CMakeLists.txt
@@ -1,0 +1,12 @@
+# add_library(libbits
+#         bits.cpp
+# )
+
+
+add_library(libendian
+        endian.cpp
+)
+
+add_library(libuarray
+        uarray.c
+)
