+++ src/GPT-originals/endian.cpp
