+++ src/GPT-originals/uarray.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <errno.h>
-#include "uarray.h"
+#include "../../include/GPT-originals/uarray.h"
