+++ src/bitvector.cpp
