+++ src/utils/CMakeLists.txt
@@ -1,0 +1,24 @@
+add_library(libdfs
+        dfs.c
+)
+
+add_library(
+        libdll
+        dll.c
+)
+
+
+
+add_library(
+        libll2
+        ll2.c
+)
+
+add_library(libqueue
+        queue.c
+)
+
+
+add_library(libstack
+        stack.c
+)
