+++ tests/CMakeLists.txt
@@ -1,0 +1,47 @@
+include_directories(${CMAKE_SOURCE_DIR}/include)
+
+# add_executable(bits-tester
+#         bits-tester.cpp
+# )
+
+# add_dependencies(bits-tester libbits)
+
+
+# add_executable(bitvector-tester
+#         bitvector-tester.cpp
+# )
+# 
+# add_dependencies(bitvector-tester libbitvector)
+# target_link_libraries(bitvector-tester libbitvector)
+
+
+
+# add_executable(dll-tester
+#         dll-tester.c
+# )
+# add_dependencies(dll-tester libdll)
+
+
+# add_executable(endian-tester
+#         endian-tester.c
+# )
+
+# add_dependencies(endian-tester libendian)
+
+
+# add_executable(jacobsonrank-tester
+#         jacobsonrank-tester.cpp
+# )
+# add_dependencies(jacobsonrank-tester libjacobson)
+
+
+# add_executable(ll2-tester
+#         ll2-tester.c
+# )
+# add_dependencies(ll2-tester libll2)
+
+
+# add_executable(uarray-tester
+#         uarray-tester.c
+# )
+# add_dependencies(uarray-tester libuarray)
