+++ tests/utils-tester.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "endian.h"
 #include <cstdio>
 #include <iostream>
 #include <bitset>
